# docs/swagger.yaml
definitions:
  api.createNoteRequest:
    properties:
      content:
        maxLength: 10000
        type: string
      title:
        maxLength: 100
        minLength: 1
        type: string
      web_ids:
        items:
          type: string
        maxItems: 5
        minItems: 1
        type: array
    required:
    - content
    - title
    type: object
  api.createUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  api.createWebRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  api.listNoteResponse:
    properties:
      notes:
        items:
          $ref: '#/definitions/api.noteResponse'
        type: array
    type: object
  api.listWebResponse:
    properties:
      webs:
        items:
          $ref: '#/definitions/api.webResponse'
        type: array
    type: object
  api.loginUserRedirectResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
      session_id:
        type: string
      user:
        $ref: '#/definitions/api.userResponse'
    required:
    - access_token
    - access_token_expires_at
    - refresh_token
    - refresh_token_expires_at
    - session_id
    - user
    type: object
  api.loginUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  api.noteResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      user_id:
        type: string
      webs:
        items:
          $ref: '#/definitions/api.webResponse'
        type: array
    type: object
  api.putNoteRequest:
    properties:
      content:
        type: string
      title:
        description: ID      string   `uri:"id" binding:"required,uuid"`
        type: string
      web_ids:
        items:
          type: string
        maxItems: 5
        minItems: 1
        type: array
    required:
    - content
    - title
    type: object
  api.registerRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  api.renewAccessTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  api.renewAccessTokenResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
    type: object
  api.userResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      password_changed_at:
        type: string
    required:
    - created_at
    - email
    - id
    - password_changed_at
    type: object
  api.webResponse:
    properties:
      created_at:
        type: string
      html:
        type: string
      id:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      url:
        type: string
      user_id:
        type: string
    required:
    - created_at
    - html
    - id
    - thumbnail_url
    - title
    - url
    - user_id
    type: object
info:
  contact: {}
paths:
  /notes:
    get:
      parameters:
      - in: query
        minimum: 1
        name: page_id
        required: true
        type: integer
      - in: query
        maximum: 10
        minimum: 5
        name: page_size
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.listNoteResponse'
      security:
      - AccessToken: []
      tags:
      - note
    post:
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createNoteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.noteResponse'
      security:
      - AccessToken: []
      tags:
      - note
  /notes/{id}:
    delete:
      parameters:
      - description: Notes ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - AccessToken: []
      tags:
      - note
    get:
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.noteResponse'
      security:
      - AccessToken: []
      tags:
      - note
    put:
      parameters:
      - description: Web ID
        in: path
        name: id
        required: true
        type: string
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.putNoteRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.noteResponse'
      security:
      - AccessToken: []
      tags:
      - note
  /register:
    post:
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.registerRequest'
      responses:
        "200":
          description: OK
          schema:
            type: ""
      tags:
      - user
  /users:
    post:
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userResponse'
      tags:
      - user
  /users/login:
    post:
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.loginUserRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.loginUserRedirectResponse'
      tags:
      - user
  /users/me:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userResponse'
      security:
      - AccessToken: []
      tags:
      - user
  /users/renew_access:
    post:
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.renewAccessTokenRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.renewAccessTokenResponse'
      tags:
      - user
  /verify:
    post:
      parameters:
      - in: query
        name: email
        required: true
        type: string
      - in: query
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userResponse'
      tags:
      - user
  /webs:
    get:
      parameters:
      - in: query
        minimum: 1
        name: page_id
        required: true
        type: integer
      - in: query
        maximum: 10
        minimum: 5
        name: page_size
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.listWebResponse'
      security:
      - AccessToken: []
      tags:
      - web
    post:
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createWebRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.webResponse'
      security:
      - AccessToken: []
      tags:
      - web
  /webs/{id}:
    delete:
      parameters:
      - description: Web ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: ""
      security:
      - AccessToken: []
      tags:
      - web
    get:
      parameters:
      - description: Web ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.webResponse'
      security:
      - AccessToken: []
      tags:
      - web
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"